+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - When dragging tasks from daily notes across days in the weekly view, they now get moved across files
 - There is now an option to hide completed tasks from timeline
+- There is now an option to hide subtasks from task blocks in the timeline
 
 ### 🐞 Fixed issues